split use case to each component
