+++ apps/anyspend-demo-nextjs/src/app/components/SwapTokensButton.tsx
@@ -1,0 +1,21 @@
+import { useModalStore } from "@b3dotfun/sdk/global-account/react";
+
+export function SwapTokensButton() {
+  const setB3ModalOpen = useModalStore(state => state.setB3ModalOpen);
+  const setB3ModalContentType = useModalStore(state => state.setB3ModalContentType);
+
+  return (
+    <button
+      onClick={() => {
+        setB3ModalOpen(true);
+        setB3ModalContentType({ type: "anySpend" });
+      }}
+      className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-blue-100 hover:shadow-md"
+    >
+      <div>
+        <h3 className="text-lg font-medium text-gray-900">Swap Tokens</h3>
+        <p className="mt-1 text-sm text-gray-500">Swap between any supported tokens instantly</p>
+      </div>
+    </button>
+  );
+}
