+++ apps/anyspend-demo-nextjs/src/app/page.tsx
@@ -1,50 +1,17 @@
 "use client";
 
-import { B3_TOKEN } from "@b3dotfun/sdk/anyspend";
-import { useModalStore } from "@b3dotfun/sdk/global-account/react";
 import { useState } from "react";
-import { base } from "viem/chains";
 import { SignInButton } from "./components/SignInButton";
 import { SignatureMintModal } from "./components/SignatureMintModal";
+import { SwapTokensButton } from "./components/SwapTokensButton";
+import { BuyWithFiatButton } from "./components/BuyWithFiatButton";
+import { MintB3kemonButton } from "./components/MintB3kemonButton";
+import { GetB3TokenButton } from "./components/GetB3TokenButton";
+import { SignatureMintButton } from "./components/SignatureMintButton";
+import { BondKitButton } from "./components/BondKitButton";
 
 export default function Home() {
-  const setB3ModalOpen = useModalStore(state => state.setB3ModalOpen);
-  const setB3ModalContentType = useModalStore(state => state.setB3ModalContentType);
   const [isSignatureMintOpen, setIsSignatureMintOpen] = useState(false);
-
-  const handleMint = async () => {
-    const usdcOnBase = {
-      chainId: base.id,
-      address: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
-      decimals: 6,
-      name: "USD Coin",
-      symbol: "USDC",
-      metadata: {
-        logoURI: "https://b3.fun/logo.png",
-      },
-    };
-    // Generate random token ID between 0 and 6
-    const randomTokenId = Math.floor(Math.random() * 7);
-
-    const nftContract = {
-      chainId: base.id,
-      contractAddress: "0xe04074c294d0Db90F0ffBC60fa61b48672C91965",
-      price: "1990000", // 1.99 USDC (6 decimals)
-      priceFormatted: "1.99",
-      currency: usdcOnBase,
-      name: "Mystery B3kemon",
-      description: "Summon a mysterious B3kemon creature!",
-      tokenId: randomTokenId,
-      type: "erc1155" as const,
-      imageUrl: null,
-    };
-
-    setB3ModalOpen(true);
-    setB3ModalContentType({
-      type: "anySpendNft",
-      nftContract,
-    });
-  };
 
   return (
     <div className="relative min-h-screen bg-[#FAFAFA]">
@@ -56,86 +23,12 @@ export default function Home() {
           <p className="mb-12 text-center text-gray-500">Experience seamless crypto transactions</p>
 
           <div className="grid grid-cols-1 gap-6 sm:grid-cols-2">
-            <button
-              onClick={() => {
-                setB3ModalOpen(true);
-                setB3ModalContentType({ type: "anySpend" });
-              }}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-blue-100 hover:shadow-md"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">Swap Tokens</h3>
-                <p className="mt-1 text-sm text-gray-500">Swap between any supported tokens instantly</p>
-              </div>
-            </button>
-
-            <button
-              onClick={() => {
-                setB3ModalOpen(true);
-                setB3ModalContentType({ type: "anySpend", defaultActiveTab: "fiat" });
-              }}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-purple-100 hover:shadow-md"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">Buy with Fiat</h3>
-                <p className="mt-1 text-sm text-gray-500">Purchase crypto directly with your credit card</p>
-              </div>
-            </button>
-
-            <button
-              onClick={handleMint}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-green-100 hover:shadow-md disabled:cursor-not-allowed disabled:opacity-50"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">Mint B3kemon</h3>
-                <p className="mt-1 text-sm text-gray-500">Mint your own mysterious B3kemon NFT</p>
-              </div>
-            </button>
-
-            <button
-              onClick={() => {
-                setB3ModalOpen(true);
-                setB3ModalContentType({
-                  type: "anySpend",
-                  destinationTokenAddress: B3_TOKEN.address,
-                  destinationTokenChainId: B3_TOKEN.chainId,
-                });
-              }}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-indigo-100 hover:shadow-md"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">Get B3 Token</h3>
-                <p className="mt-1 text-sm text-gray-500">Purchase B3 tokens directly through AnySpend</p>
-              </div>
-            </button>
-
-            <button
-              onClick={() => setIsSignatureMintOpen(true)}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-yellow-100 hover:shadow-md"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">Signature Mint</h3>
-                <p className="mt-1 text-sm text-gray-500">Mint NFTs using signature-based minting</p>
-              </div>
-            </button>
-            <button
-              onClick={() => {
-                setB3ModalOpen(true);
-                setB3ModalContentType({
-                  type: "anySpendBondKit",
-                  recipientAddress: "0xd67C2dfb4862Ae6F5c079FbEbF403dC01E5Df3f0",
-                  contractAddress: "0x01ef4b21bb06fa40fe08e112a472b8925bb271d1",
-                  ethAmount: "0.001",
-                  minTokensOut: "0",
-                });
-              }}
-              className="group flex h-40 flex-col justify-between overflow-hidden rounded-lg border border-gray-100 bg-white p-6 text-left shadow-sm transition-all hover:border-blue-100 hover:shadow-md"
-            >
-              <div>
-                <h3 className="text-lg font-medium text-gray-900">BondKit</h3>
-                <p className="mt-1 text-sm text-gray-500">Buy BondKit tokens with ETH</p>
-              </div>
-            </button>
+            <SwapTokensButton />
+            <BuyWithFiatButton />
+            <MintB3kemonButton />
+            <GetB3TokenButton />
+            <SignatureMintButton onClick={() => setIsSignatureMintOpen(true)} />
+            <BondKitButton />
           </div>
         </div>
       </div>
